WIP. Started implementing geoservice helpers
