+++ viewer-admin/pom.xml
@@ -133,6 +133,10 @@
             <groupId>org.mockito</groupId>
             <artifactId>mockito-core</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.flamingo-mc</groupId>
+            <artifactId>viewer-config-persistence-helpers</artifactId>
+        </dependency>
     </dependencies>
     <build>
         <plugins>
