+++ viewer-config-persistence-helpers/src/main/java/nl/b3p/viewer/config/services/UpdateResult.java
@@ -23,6 +23,7 @@ import java.util.Map;
 import java.util.SortedMap;
 import java.util.TreeMap;
 import javax.persistence.EntityManager;
+
 import nl.b3p.web.WaitPageStatus;
 import org.apache.commons.lang3.mutable.MutableInt;
 import org.apache.commons.lang3.tuple.MutablePair;
