+++ viewer-config-persistence/src/main/java/nl/b3p/viewer/config/services/Updatable.java
@@ -24,5 +24,5 @@ import javax.persistence.EntityManager;
  */
 public interface Updatable {
     
-    public UpdateResult update(EntityManager em);
+    public void update(EntityManager em);
 }
